requirements.txt
asgiref==3.8.1
coverage==7.7.1
Django==5.1.7
django-filter==25.1
djangorestframework==3.15.2
sqlparse==0.5.3
